+++ packages/safe-ethers-adapters/src/index.ts
@@ -1,1 +1,2 @@
+export * from './service'
 export * from './signer'
